Revert "some fixes"
This reverts commit e86bade501297f4f6815fb77fbdad3322db27554.

+++ PathFind/ValueRange.Tests/InclusiveValueRangeTests.cs
@@ -121,7 +121,7 @@ namespace Common.Tests
         [Test]
         public void Amplitude_ExtremeValuesOfInt_ReturnsUintMaxValue()
         {
-            long amplitude = ValueRanges.IntRange.Amplitude();
+            uint amplitude = ValueRanges.IntRange.Amplitude();
 
             Assert.AreEqual(amplitude, uint.MaxValue);
         }
